# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: >-
    Keep in touch with your friends
    by sharing photos of special moments,
    thanks to WASAPhoto!

    You can upload your photos directly from your PC,
    and they will be visible to everyone following you.
  version: 0.0.1
...
components:
  schemas:
    UserID:
      description: UserID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    PhotoID:
      description: PhotoID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    CommentID:
      description: CommentID.
      readOnly: true
      type: number
      format: uint64
      example: 1
    Username:
      description: Username.
      type: string
      pattern: '^.*?$'
      minLength: 3
      maxLength: 16
      example: Maria
    User:
      description: User object.
      type: object
      properties:
        userID:
          $ref: '#/components/schemas/UserID'
        username:
          $ref: '#/components/schemas/Username'
    Date:
      description: Publish date.
      type: string
      format: date-time
      minLength: 19
      maxLength: 19
      example: '2006-01-02 15:04:05'
    Photo:
      description: Photo object.
      type: object
      properties:
        photoID:
          $ref: '#/components/schemas/PhotoID'
        authorID:
          $ref: '#/components/schemas/UserID'
        format:
          description: Photo format.
          type: string
          enum:
            - png
            - jpg
          pattern: '^.*?$'
          minLength: 3
          maxLength: 3
          example: png
        date:
          $ref: '#/components/schemas/Date'
        likesList:
          description: Users list.
          type: array
          minItems: 0
          maxItems: 99
          items: { $ref: '#/components/schemas/User' }
        commentsList:
          description: Comments list.
          type: array
          minItems: 0
          maxItems: 99
          items: { $ref: '#/components/schemas/Comment' }
    CommentText:
          description: Comment text.
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2200
          example: Amazing photo!
    Comment:
      description: Comment object.
      type: object
      properties:
        commentID:
          $ref: '#/components/schemas/CommentID'
        photoID:
          $ref: '#/components/schemas/PhotoID'
        authorID:
          $ref: '#/components/schemas/UserID'
        commentText:
          $ref: '#/components/schemas/CommentText'
        date:
          $ref: '#/components/schemas/Date'
    Profile:
      description: Profile object.
      type: object
      properties:
        username:
          $ref: '#/components/schemas/Username'
        photosCount:
          description: Total number of photos.
          type: number
          format: uint64
          example: 53
        followersCount:
          description: Total number of followers.
          type: number
          format: uint64
          example: 53
        followingCount:
          description: Total number of following.
          type: number
          format: uint64
          example: 53
        isItMe:
          description: Is it me.
          type: boolean
          example: true
        doIFollowUser:
          description: Do I follow user.
          type: boolean
          example: true
        isInMyBannedList:
          description: Is in my banned list.
          type: boolean
          example: true
        amIBanned:
          description: Am I banned.
          type: boolean
          example: true
    BinaryImage:
      description: Binary image.
      type: string
      format: binary
      minLength: 4
      maxLength: 65535
      example: iVBORw0KGgoAAAANSU...
  parameters:
    UserID:
      name: uid
      in: path
      required: true
      description: UserID.
      schema:
        $ref: '#/components/schemas/UserID'
    PhotoID:
      name: pid
      in: path
      required: true
      description: PhotoID.
      schema:
        $ref: '#/components/schemas/PhotoID'
    CommentID:
      name: cid
      in: path
      required: true
      description: CommentID.
      schema:
        $ref: '#/components/schemas/CommentID'
...
  responses:
    Unauthorized:
      # 401
      description: The client must authenticate.
    Forbidden:
      # 403
      description: Some user has banned the other.
    InternalServerError:
      #500
      description: Server error.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
...
  /photos/{pid}/:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: delete a photo
      description: Delete previously uploaded photo.
      responses:
        "204":
          # No Content
          description: User deleted photo.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          description: Requesting user not author of photo.
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    get:
      operationId: getPhoto
      tags: [getters]
      summary: get binary image
      description: Get binary image.
      responses:
        "200":
          # OK
          description: Photo file retrieved successfully.
          content:
            image/*:
              schema: { $ref: '#/components/schemas/BinaryImage' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
  /likes/{pid}:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    put:
      operationId: likePhoto
      tags: [likes]
      summary: put a like to a photo
      description: Put a like to a photo.
      responses:
        "204":
          # No Content
          description: The like has been put to the specified photo.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    delete:

      operationId: unlikePhoto
      tags: [likes]
      summary: unlike a photo
      description: Unlike photo.
      responses:
        "204":
          # No Content
          description: The like has been removed from the specified photo.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /photos/{pid}/comments/:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    post:
      operationId: commentPhoto
      tags: [comments]
      summary: comment on a photo
      description: Comment on a photo.
      requestBody:
        description: Comment text.
        required: true
        content:
          text/plain:
            schema: { $ref: '#/components/schemas/CommentText' }
      responses:
        "201":
          # Created
          description: Comment successfully added to the specified photo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo or;
            invalid comment data.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'

    get:
      operationId: getCommentsList
      tags: [getters]
      summary: get photo comments
      description: Get the list of comments for a specific photo.
      responses:
        "200":
          # OK
          description: List of photo comments retrieved successfully.
          content:
            application/json:
              schema:
                description: Comments list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Comment' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /photos/{pid}/comments/{cid}:

    parameters:
      - $ref: '#/components/parameters/PhotoID'
      - $ref: '#/components/parameters/CommentID'

    delete:
      operationId: uncommentPhoto
      tags: [comments]
      summary: remove a comment from a photo
      description: Remove a comment from a photo.
      responses:
        "204":
          # No Content
          description: Comment successfully removed from the specified photo.
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo or;
            invalid path parameter cid or;
            cid not matching any existing comment.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          description: Requesting user not author of comment.
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /users/{uid}/:

    parameters:
      - $ref: '#/components/parameters/UserID'

    get:
      operationId: getUserProfile
      tags: [getters]
      summary: get user profile
      description: Get user profile by user ID.
      responses:
        "200":
          # OK
          description: User profile retrieved successfully.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
  /users/{uid}/photos/:

    parameters:
      - $ref: '#/components/parameters/UserID'

    get:
      operationId: getPhotosList
      tags: [getters]
      summary: get user photos
      description: Get the list of photos uploaded by a user.
      responses:
        "200":
          # OK
          description: List of user photos retrieved successfully.
          content:
            application/json:
              schema:
                description: Photos list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/Photo' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter uid or;
            uid not matching any existing user.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{pid}/likes/:

    parameters:
      - $ref: '#/components/parameters/PhotoID'

    get:
      operationId: getLikesList
      tags: [getters]
      summary: get photo likes
      description: Get the list of users who liked a specific photo.
      responses:
        "200":
          # OK
          description: List of photo likes retrieved successfully.
          content:
            application/json:
              schema:
                description: Users list.
                type: array
                minItems: 0
                maxItems: 99
                items: { $ref: '#/components/schemas/User' }
        "400":
          # Bad Request
          description: |-
            Invalid path parameter pid or;
            pid not matching any existing photo.
        "401":
          # Unauthorized
          $ref: '#/components/responses/Unauthorized'
        "403":
          # Forbidden
          $ref: '#/components/responses/Forbidden'
        "500":
          # Internal Server Error
          $ref: '#/components/responses/InternalServerError'
...